# src/main/webapp/app/shared/api/generated/API-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "OncoKB™ is a precision oncology knowledge base developed at Memorial Sloan Kettering Cancer Center that contains biological and clinical information about genomic alterations in cancer.",
    "version": "v1.0.0",
    "title": "OncoKB™ APIs",
    "contact": {
      "name": "OncoKB™",
      "url": "https://www.oncokb.org",
      "email": "contact@oncokb.org"
    },
    "license": {
      "name": "Terms of Use",
      "url": "https://www.oncokb.org/terms"
    }
  },
...
  "paths": {
    "/api/account": {
      "get": {
        "tags": [
          "account-resource"
        ],
        "summary": "getAccount",
        "operationId": "getAccountUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "saveAccount",
        "operationId": "saveAccountUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userDTO",
            "description": "userDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/active-trial/finish": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "finishTrialAccountActivation",
        "operationId": "finishTrialAccountActivationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "keyAndTermsVM",
            "description": "keyAndTermsVM",
            "required": true,
            "schema": {
              "$ref": "#/definitions/KeyAndTermsVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/active-trial/info": {
      "get": {
        "tags": [
          "account-resource"
        ],
        "summary": "getTrialAccountActivationInfo",
        "operationId": "getTrialAccountActivationInfoUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/active-trial/init": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "initiateTrialAccountActivation",
        "operationId": "initiateTrialAccountActivationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "login",
            "description": "login",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/change-password": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "changePassword",
        "operationId": "changePasswordUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "passwordChangeDto",
            "description": "passwordChangeDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PasswordChangeDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/generate-reset-key": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "generateResetKey",
        "operationId": "generateResetKeyUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "login",
            "description": "login",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/request-api-access": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "requestApiAccess",
        "operationId": "requestApiAccessUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "apiAccessRequest",
            "description": "apiAccessRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApiAccessRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/resend-verification": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "resendVerification",
        "operationId": "resendVerificationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "loginVM",
            "description": "loginVM",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/account/reset-password/finish": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "finishPasswordReset",
        "operationId": "finishPasswordResetUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "keyAndPassword",
            "description": "keyAndPassword",
            "required": true,
            "schema": {
              "$ref": "#/definitions/KeyAndPasswordVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "ApiAccessRequest": {
      "type": "object",
      "properties": {
        "justification": {
          "type": "string"
        },
        "requested": {
          "type": "boolean"
        }
      },
      "title": "ApiAccessRequest"
    },
...
    "KeyAndPasswordVM": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      },
      "title": "KeyAndPasswordVM"
    },
    "KeyAndTermsVM": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "readAndAgreeWithTheTerms": {
          "type": "boolean"
        }
      },
      "title": "KeyAndTermsVM"
    },
...
    "LoginVM": {
      "type": "object",
      "required": [
        "password",
        "username"
      ],
      "properties": {
        "password": {
          "type": "string",
          "minLength": 4,
          "maxLength": 100
        },
        "rememberMe": {
          "type": "boolean"
        },
        "username": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50
        }
      },
      "title": "LoginVM"
    },
...
    "PasswordChangeDTO": {
      "type": "object",
      "properties": {
        "currentPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      },
      "title": "PasswordChangeDTO"
    },
...
  }
}